+++ test/test_imap_compare.py
@@ -18,7 +18,7 @@ import pytest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 
-import compare_imap_folders
+import imap_compare as compare_imap_folders
 from conftest import temp_argv, temp_env
