+++ test/test_imap_count.py
@@ -16,8 +16,8 @@ import pytest
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 
-import count_imap_emails
 import imap_common
+import imap_count as count_imap_emails
 from conftest import temp_argv, temp_env
