+++ test/test_migrate_resume.py
@@ -6,7 +6,7 @@ import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 
-import migrate_imap_emails
+import imap_migrate as migrate_imap_emails
 import restore_cache
 from conftest import temp_argv, temp_env
